Split files
Extract beghilosz encoder in a separate file
Add tests for Beghilosz function
