+++ beghilosz_test.go
@@ -1,0 +1,29 @@
+package encodor
+
+import (
+	"testing"
+)
+
+func TestBeghilosz(t *testing.T) {
+	type args struct {
+		input string
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{name: "Basic", args: args{"beghilosz"}, want: "250714638"},
+		{name: "Two words", args: args{"two words"}, want: "5DR0W 0WT"},
+		{name: "Hashtags untouched, but capitalized", args: args{"#hashtag"}, want: "#HASHTAG"},
+		{name: "Usernames untouched, but capitalized", args: args{"@username"}, want: "@USERNAME"},
+		{name: "Word order in a sentence with hashtag", args: args{"Sentence with a #hashtag"}, want: "#HASHTAG A 4T1W 3CN3TN35"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := Beghilosz(tt.args.input); got != tt.want {
+				t.Errorf("Beghilosz() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
