+++ go.mod
@@ -1,0 +1,3 @@
+module encodor
+
+go 1.13
